Yttr-InRed Loot Table Dump (!1)
These are as basic as loot tables can get, quick work

While testing I discovered that the Demo Cycler instantly breaks when punched; it also doesn't require a tool to harvest.
I added a loot table for it anyway, but I'm not sure if it's meant to be available outside of Creative Mode in the first place.

Co-authored-by: Kat <Kat@sleeping.town>
Reviewed-on: https://git.sleeping.town/unascribed/Yttr/pulls/1
Co-authored-by: Kat <kat@noreply@sleeping.town>
Co-committed-by: Kat <kat@noreply@sleeping.town>