Completely redo slope smoothing via GJK collision detection
Nearly perfect for axis-aligned slopes, needs work still on steep slopes and corner slopes.
Good enough for now — better than it was, certainly…

In order to fix the bad cases, I think capsule collision needs to be added, and that's a whole thing.

Also, could use adjusting player movement to align with slopes, instead of trying to fudge it while they're midair.
That could be… complicated.

Sure doesn't look like it, but this represents 6 hours of work. It is 4 AM. Please help